backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
pdfplumber==0.11.0
google-generativeai==0.8.3
pydantic==2.6.3
python-dotenv==1.0.1
python-multipart==0.0.9
